pyproject.toml
[tool.poetry]
name = "quiz_bot"
version = "0.1.0"
description = ""
authors = ["aevtikheev <eft000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "13.1"
environs = "9.2.0"
vk_api = "11.9.1"
redis = "3.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
